% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yamlpack.R
\name{.write_yamlpack_l}
\alias{.write_yamlpack_l}
\title{Write the list to a yaml file}
\usage{
.write_yamlpack_l(yamlpack_l, file_path)
}
\arguments{
\item{yamlpack_l}{a list of packages}

\item{file_path}{file path to save}
}
\description{
Write the list to a yaml file
}
\keyword{internal}
