% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_yamlpack.R
\name{install_yamlpack}
\alias{install_yamlpack}
\title{Install Packages From a yamlpack YAML File}
\usage{
install_yamlpack(
  yamlpack,
  repo = c("CRAN", "Bioconductor", "GitHub"),
  method = "pak",
  cran_url = "https://cran.ma.imperial.ac.uk/"
)
}
\arguments{
\item{yamlpack}{A yamlpack list read in with `read_yamlpack()`}

\item{repo}{one or more repos from the yamlpack to install (default all)}

\item{method}{Use conventional install or parallelize with the `pak` package}
}
\description{
Install Packages From a yamlpack YAML File
}
