% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yamlpack.R
\name{read_yamlpack}
\alias{read_yamlpack}
\title{Read a Previously Saved yamlpack File}
\usage{
read_yamlpack(file_path)
}
\arguments{
\item{file_path}{Path to a yaml file previous saved with `write_yamlpack()`}
}
\value{
yml a YAML object
}
\description{
Read a Previously Saved yamlpack File
}
