% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yamlpack.R
\name{write_yamlpack}
\alias{write_yamlpack}
\title{Write Installed Packages And Sources To A yamlpack File}
\usage{
write_yamlpack(
  file_path = file.path(getwd(), "yamlpack.yml"),
  overwrite = TRUE
)
}
\arguments{
\item{file_path}{Path to save the yamlpack file}
}
\description{
Write Installed Packages And Sources To A yamlpack File
}
